consolidate warning classes into _warnings module
